devide files
