+++ go/auth.go
@@ -1,0 +1,103 @@
+package main
+
+import (
+	"context"
+	"io"
+	"os"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/s3"
+	"github.com/aws/aws-sdk-go-v2/service/ssm"
+	"golang.org/x/oauth2"
+)
+
+type Instances struct {
+	SSMClient *ssm.Client
+	S3Client  *s3.Client
+}
+
+func (instances *Instances) getParameter(parameterName string) (*string, error) {
+	getParameterOutput, err := instances.SSMClient.GetParameter(context.TODO(), &ssm.GetParameterInput{
+		Name:           aws.String(parameterName),
+		WithDecryption: aws.Bool(true),
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return getParameterOutput.Parameter.Value, nil
+}
+
+func (instances *Instances) getRefreshToken() (*string, error) {
+	getObjectOutput, err := instances.S3Client.GetObject(context.TODO(), &s3.GetObjectInput{
+		Bucket: refreshCbBucketName,
+		Key:    refreshCbFileName,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	defer getObjectOutput.Body.Close()
+	body, err := io.ReadAll(getObjectOutput.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	refreshToken := string(body)
+
+	return &refreshToken, nil
+}
+
+func getFitbitConfig(clientID string, clientSecret string) *oauth2.Config {
+	return &oauth2.Config{
+		ClientID:     clientID,
+		ClientSecret: clientSecret,
+		Endpoint: oauth2.Endpoint{
+			TokenURL: "https://api.fitbit.com/oauth2/token",
+		},
+	}
+}
+
+func (instances *Instances) refreshAccessToken(ctx context.Context, clientID string, clientSecret string, refreshToken string) (*string, error) {
+	config := getFitbitConfig(clientID, clientSecret)
+	token := &oauth2.Token{
+		RefreshToken: refreshToken,
+	}
+
+	newToken, err := config.TokenSource(ctx, token).Token()
+	if err != nil {
+		return nil, err
+	}
+
+	tempFileName := "/tmp/" + *refreshCbBucketName
+
+	tempFile, err := os.Create(tempFileName)
+	if err != nil {
+		return nil, err
+	}
+	defer tempFile.Close()
+
+	newRefreshToken := newToken.RefreshToken
+	data := []byte(newRefreshToken)
+	_, err = tempFile.Write(data)
+	if err != nil {
+		return nil, err
+	}
+
+	file, err := os.Open(tempFileName)
+	if err != nil {
+		return nil, err
+	}
+	defer file.Close()
+
+	_, err = instances.S3Client.PutObject(context.TODO(), &s3.PutObjectInput{
+		Bucket: refreshCbBucketName,
+		Key:    refreshCbFileName,
+		Body:   file,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return &newToken.AccessToken, nil
+}
